requirements.txt
# requirements.txt
mysql-connector-python==9.2.0
beautifulsoup4==4.11.1
requests==2.28.1
nltk==3.7
spacy==3.8.3
scikit-learn==1.3.2
torch==2.6.0
transformers==4.48.3
pandas==1.5.3
numpy==1.24.4
flask==2.0.1
python-Levenshtein==0.26.1
fuzzywuzzy==0.18.0
-e .